Use uint64_t to ensure compiler compatibility

+++ src/fstcore/interface/ifstcolumn.h
@@ -143,7 +143,7 @@ public:
   virtual void BufferToVec(uint64_t nrOfElements, uint64_t startElem, uint64_t endElem,
 	  uint64_t vecOffset, unsigned int* sizeMeta, char* buf) = 0;
 
-  virtual const char* GetElement(unsigned long long elementNr) = 0;
+  virtual const char* GetElement(uint64_t elementNr) = 0;
 };
